# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dot.Agents API
  version: 1.0.0
  description: |
    Enterprise AI Workforce Platform API.
    All endpoints require `auth:sanctum` + `org.context` middleware.
    Responses are scoped to the authenticated user's current organization.
  contact:
    name: Dot.Agents Platform
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Production API v1
security:
  - bearerToken: []
paths:
...
  /agents:
    get:
      summary: List agent marketplace catalog
      operationId: listAgents
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: department
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: agent_type
          in: query
          schema:
            type: string
            enum: [advisory, semi-autonomous, autonomous, executive_approval]
      responses:
        '200':
          description: Paginated agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAgents'
  /agents/{agent}:
    get:
      summary: Get agent details
      operationId: showAgent
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/AgentParam'
      responses:
        '200':
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Deployments ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SearchParam:
      name: search
      in: query
      schema:
        type: string
        maxLength: 100
    AgentParam:
      name: agent
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
        agent_type:
          type: string
          enum: [advisory, semi-autonomous, autonomous, executive_approval]
        status:
          type: string
          enum: [active, beta, deprecated, archived]
        is_enterprise_only:
          type: boolean
        base_price:
          type: number
          format: decimal
        certification_score:
          type: integer
          minimum: 0
          maximum: 100
        trust_tier:
          type: string
          enum: [uncertified, bronze, silver, gold, platinum]
        certified_at:
          type: string
          format: date-time
          nullable: true
...
    PaginatedAgents:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...